pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = ["Volodymyr Vovnenko <vovnenko.volodymyr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
uvicorn = "^0.32.1"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pydantic-settings = "^2.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"
cloudinary = "^1.41.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
